requirements_azure.txt
# Requirements for Azure App Service deployment
# Generated on 2025-09-17T04:27:00.573768

# Base application requirements
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3
python-dotenv==1.0.1
requests==2.32.3
authlib==1.3.1
pandas==2.2.2
openai==1.35.15
gunicorn==22.0.0
pyodbc==5.1.0
pymssql==2.3.1
email-validator==2.1.1
azure-communication-email==1.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
bleach==6.1.0

# Azure-specific requirements
azure-communication-email>=1.0.0
azure-messaging-webpubsubservice>=1.0.0
azure-servicebus>=7.0.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
redis>=4.5.0
Flask-Session>=0.5.0
applicationinsights>=0.11.10
gunicorn>=20.1.0
gevent>=22.10.0
psycopg2-binary>=2.9.0